+++ packages/block-editor/src/components/global-styles/typography-panel.js
@@ -38,7 +38,12 @@ import {
 	findNearestStyleAndWeight,
 } from './typography-utils';
 import { getFontStylesAndWeights } from '../../utils/get-font-styles-and-weights';
-import { getInheritanceProps, InheritanceToolsPanelItem } from './inheritance';
+import {
+	getCommonInheritanceTooltipText,
+	getInheritanceProps,
+	getInheritanceTooltipTextByPath,
+	InheritanceToolsPanelItem,
+} from './inheritance';
 
 const MIN_TEXT_COLUMNS = 1;
 const MAX_TEXT_COLUMNS = 6;
@@ -221,6 +226,7 @@ export default function TypographyPanel( {
 	value,
 	onChange,
 	inheritedValue = value,
+	inheritedSources = {},
 	settings,
 	panelId,
 	defaultControls = DEFAULT_CONTROLS,
@@ -234,6 +240,10 @@ export default function TypographyPanel( {
 		showInheritanceLabelIndicators
 			? getInheritanceProps( isInherited, hasLocalOverride, className )
 			: {};
+	const tooltipText = ( path ) =>
+		getInheritanceTooltipTextByPath( inheritedSources, path );
+	const commonTooltipText = ( paths ) =>
+		getCommonInheritanceTooltipText( inheritedSources, paths );
 
 	// Text color. Writes to `color.text` (unchanged storage path). The
 	// control is rendered here instead of the Color panel because text
@@ -795,6 +805,7 @@ export default function TypographyPanel( {
 					showInheritanceLabelIndicators={
 						showInheritanceLabelIndicators
 					}
+					inheritedSources={ inheritedSources }
 					isPlaceholder={
 						userTextColor === undefined && textColor !== undefined
 					}
@@ -803,6 +814,7 @@ export default function TypographyPanel( {
 						{
 							key: 'text',
 							label: __( 'Color' ),
+							sourcePaths: [ 'color.text' ],
 							inheritedValue: textColor,
 							inheritedSlug:
 								extractPresetSlug(
@@ -834,6 +846,9 @@ export default function TypographyPanel( {
 						hasFontFamily() && inheritedFontFamily !== undefined
 					) }
 					label={ __( 'Font' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.fontFamily'
+					) }
 					hasValue={ hasFontFamily }
 					onDeselect={ resetFontFamily }
 					isShownByDefault={ defaultControls.fontFamily }
@@ -853,6 +868,9 @@ export default function TypographyPanel( {
 						hasFontSize() && rawInheritedFontSize !== undefined
 					) }
 					label={ __( 'Size' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.fontSize'
+					) }
 					hasValue={ hasFontSize }
 					hasInlineEndToggle
 					onDeselect={ resetFontSize }
@@ -879,6 +897,10 @@ export default function TypographyPanel( {
 								inheritedFontWeight !== undefined )
 					) }
 					label={ appearanceControlLabel }
+					inheritanceTooltipText={ commonTooltipText( [
+						'typography.fontStyle',
+						'typography.fontWeight',
+					] ) }
 					hasValue={ hasFontAppearance }
 					onDeselect={ resetFontAppearance }
 					isShownByDefault={ defaultControls.fontAppearance }
@@ -904,6 +926,9 @@ export default function TypographyPanel( {
 						'single-column'
 					) }
 					label={ __( 'Line height' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.lineHeight'
+					) }
 					hasValue={ hasLineHeight }
 					onDeselect={ resetLineHeight }
 					isShownByDefault={ defaultControls.lineHeight }
@@ -936,6 +961,9 @@ export default function TypographyPanel( {
 						'single-column'
 					) }
 					label={ __( 'Letter spacing' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.letterSpacing'
+					) }
 					hasValue={ hasLetterSpacing }
 					onDeselect={ resetLetterSpacing }
 					isShownByDefault={ defaultControls.letterSpacing }
@@ -969,6 +997,9 @@ export default function TypographyPanel( {
 						hasTextIndent() && inheritedTextIndent !== undefined
 					) }
 					label={ __( 'Line indent' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.textIndent'
+					) }
 					hasValue={ hasTextIndent }
 					onDeselect={ resetTextIndent }
 					isShownByDefault={ defaultControls.textIndent }
@@ -1010,6 +1041,9 @@ export default function TypographyPanel( {
 						'single-column'
 					) }
 					label={ __( 'Columns' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.textColumns'
+					) }
 					hasValue={ hasTextColumns }
 					onDeselect={ resetTextColumns }
 					isShownByDefault={ defaultControls.textColumns }
@@ -1040,6 +1074,9 @@ export default function TypographyPanel( {
 						'single-column'
 					) }
 					label={ __( 'Decoration' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.textDecoration'
+					) }
 					hasValue={ hasTextDecoration }
 					onDeselect={ resetTextDecoration }
 					isShownByDefault={ defaultControls.textDecoration }
@@ -1061,6 +1098,9 @@ export default function TypographyPanel( {
 						'single-column'
 					) }
 					label={ __( 'Orientation' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.writingMode'
+					) }
 					hasValue={ hasWritingMode }
 					onDeselect={ resetWritingMode }
 					isShownByDefault={ defaultControls.writingMode }
@@ -1081,6 +1121,9 @@ export default function TypographyPanel( {
 							inheritedTextTransform !== undefined
 					) }
 					label={ __( 'Letter case' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.textTransform'
+					) }
 					hasValue={ hasTextTransform }
 					onDeselect={ resetTextTransform }
 					isShownByDefault={ defaultControls.textTransform }
@@ -1102,6 +1145,9 @@ export default function TypographyPanel( {
 						hasTextAlign() && inheritedTextAlign !== undefined
 					) }
 					label={ __( 'Text alignment' ) }
+					inheritanceTooltipText={ tooltipText(
+						'typography.textAlign'
+					) }
 					hasValue={ hasTextAlign }
 					onDeselect={ resetTextAlign }
 					isShownByDefault={ defaultControls.textAlign }