+++ packages/block-editor/src/hooks/border.js
@@ -170,11 +170,8 @@ export function BorderPanel( { clientId, name, setAttributes, settings } ) {
 		[ clientId, isEnabled ]
 	);
 
-	const { value: inheritedValue } = useResolvedStyles(
-		name,
-		className,
-		selectedState
-	);
+	const { value: inheritedValue, sources: inheritedSources } =
+		useResolvedStyles( name, className, selectedState );
 
 	const isStateSelected = ! isDefaultBlockStyleState( selectedState );
 
@@ -219,6 +216,7 @@ export function BorderPanel( { clientId, name, setAttributes, settings } ) {
 			onChange={ onChange }
 			defaultControls={ defaultControls }
 			inheritedValue={ inheritedValue }
+			inheritedSources={ inheritedSources }
 		/>
 	);
 }