+++ packages/block-editor/src/hooks/dimensions.js
@@ -95,11 +95,8 @@ export function DimensionsPanel( { clientId, name, setAttributes, settings } ) {
 		[ clientId, isEnabled ]
 	);
 
-	const { value: inheritedValue } = useResolvedStyles(
-		name,
-		className,
-		selectedState
-	);
+	const { value: inheritedValue, sources: inheritedSources } =
+		useResolvedStyles( name, className, selectedState );
 
 	const [ visualizedProperty, setVisualizedProperty ] = useVisualizer();
 	const value = isStateSelected
@@ -152,6 +149,7 @@ export function DimensionsPanel( { clientId, name, setAttributes, settings } ) {
 					isStateSelected ? undefined : setVisualizedProperty
 				}
 				inheritedValue={ inheritedValue }
+				inheritedSources={ inheritedSources }
 			/>
 			{ ! isStateSelected &&
 				!! settings?.spacing?.padding &&